support for custom variables & refactoring stuff out off grizzly.types
clean up of code that loads variables, which resulted in
refactoring of grizzly.types.

since they are more than just simple types, and should be closer to
their functionality.
