+++ grizzly/testdata/utils.py
@@ -1,6 +1,6 @@
 import logging
 
-from typing import Optional, List, Dict, Any, Tuple, Set, cast
+from typing import TYPE_CHECKING, Optional, List, Dict, Any, Tuple, Set, cast
 from collections import namedtuple
 from os import environ
 from time import perf_counter as time
@@ -9,17 +9,21 @@ from jinja2 import Template, Environment
 from jinja2.meta import find_undeclared_variables
 from locust.exception import StopUser
 
-from ..context import GrizzlyContext
 from ..tasks import GrizzlyTask
-from ..types import RequestType, TestdataType, GrizzlyDictValueType, GrizzlyDict
+from ..types import RequestType, TestdataType, GrizzlyDictValueType
 from ..utils import merge_dicts
 from .ast import get_template_variables
+from . import GrizzlyDict
+
+
+if TYPE_CHECKING:
+    from ..context import GrizzlyContext
 
 
 logger = logging.getLogger(__name__)
 
 
-def initialize_testdata(tasks: List[GrizzlyTask]) -> Tuple[TestdataType, Set[str]]:
+def initialize_testdata(grizzly: 'GrizzlyContext', tasks: List[GrizzlyTask]) -> Tuple[TestdataType, Set[str]]:
     testdata: TestdataType = {}
     template_variables = get_template_variables(tasks)
 
@@ -30,13 +34,16 @@ def initialize_testdata(tasks: List[GrizzlyTask]) -> Tuple[TestdataType, Set[str
         testdata[scenario] = {}
 
         for variable in variables:
-            if variable.count('.') > 1:
-                variable_datatype = '.'.join(variable.split('.')[0:2])
+            module_name, variable_type, variable_name = GrizzlyDict.get_variable_spec(variable)
+            if module_name is not None and variable_type is not None:
+                variable_datatype = f'{variable_type}.{variable_name}'
+                if module_name != 'grizzly.testdata.variables':
+                    variable_datatype = f'{module_name}.{variable_datatype}'
             else:
-                variable_datatype = variable
+                variable_datatype = variable_name
 
             if variable_datatype not in initialized_datatypes:
-                initialized_datatypes[variable_datatype], dependencies = _get_variable_value(variable_datatype)
+                initialized_datatypes[variable_datatype], dependencies = _get_variable_value(grizzly, variable_datatype)
                 external_dependencies.update(dependencies)
 
             testdata[scenario][variable] = initialized_datatypes[variable_datatype]
@@ -44,14 +51,14 @@ def initialize_testdata(tasks: List[GrizzlyTask]) -> Tuple[TestdataType, Set[str
     return testdata, external_dependencies
 
 
-def _get_variable_value(name: str) -> Tuple[Any, Set[str]]:
-    grizzly = GrizzlyContext()
-    default_value = grizzly.state.variables.get(name, None)
+def _get_variable_value(grizzly: 'GrizzlyContext', name: str) -> Tuple[Any, Set[str]]:
     external_dependencies: Set[str] = set()
 
-    if '.' in name:
-        variable_type, variable_name = name.rsplit('.', 1)
-        variable = GrizzlyDict.load_variable(variable_type)
+    default_value = grizzly.state.variables.get(name, None)
+    module_name, variable_type, variable_name = GrizzlyDict.get_variable_spec(name)
+
+    if module_name is not None and variable_type is not None:
+        variable = GrizzlyDict.load_variable(module_name, variable_type)
         external_dependencies = variable.__dependencies__
         if getattr(variable, '__on_consumer__', False):
             value = cast(Any, '__on_consumer__')
@@ -66,46 +73,42 @@ def _get_variable_value(name: str) -> Tuple[Any, Set[str]]:
     return value, external_dependencies
 
 
-def transform(data: Dict[str, Any], objectify: Optional[bool] = True) -> Dict[str, Any]:
+def transform(grizzly: 'GrizzlyContext', data: Dict[str, Any], objectify: Optional[bool] = True) -> Dict[str, Any]:
     testdata: Dict[str, Any] = {}
 
     for key, value in data.items():
+        module_name, variable_type, variable_name = GrizzlyDict.get_variable_spec(key)
+
         if '.' in key:
-            if value == '__on_consumer__':
-                grizzly = GrizzlyContext()
-                variable_type, variable_name = key.rsplit('.', 1)
-                if 'Custom' in key:
-                    print(f'transform: {variable_type=}, {variable_name=}')
-                variable_class_type = GrizzlyDict.load_variable(variable_type)
-                try:
+            if module_name is not None and variable_type is not None:
+                if value == '__on_consumer__':
+                    if 'Custom' in key:
+                        print(f'transform: {module_name=}, {variable_type=}, {variable_name=}')
+
+                    variable_type_instance = GrizzlyDict.load_variable(module_name, variable_type)
                     initial_value = grizzly.state.variables.get(key, None)
-                    variable_instance = variable_class_type(variable_name, initial_value)
-                except ValueError as e:
-                    if 'object already has attribute' in str(e):
-                        variable_instance = variable_class_type.get()
-                    else:
-                        raise e
+                    variable_instance = variable_type_instance.obtain(variable_name, initial_value)
 
-                start_time = time()
-                exception: Optional[Exception] = None
-                try:
-                    value = variable_instance[variable_name]
-                except Exception as e:
-                    exception = e
-                    logger.error(str(e), exc_info=grizzly.state.verbose)
-                finally:
-                    response_time = int((time() - start_time) * 1000)
-                    grizzly.state.locust.environment.events.request.fire(
-                        request_type=RequestType.VARIABLE(),
-                        name=key,
-                        response_time=response_time,
-                        response_length=0,
-                        context=None,
-                        exception=exception,
-                    )
+                    start_time = time()
+                    exception: Optional[Exception] = None
+                    try:
+                        value = variable_instance[variable_name]
+                    except Exception as e:
+                        exception = e
+                        logger.error(str(e), exc_info=grizzly.state.verbose)
+                    finally:
+                        response_time = int((time() - start_time) * 1000)
+                        grizzly.state.locust.environment.events.request.fire(
+                            request_type=RequestType.VARIABLE(),
+                            name=key,
+                            response_time=response_time,
+                            response_length=0,
+                            context=None,
+                            exception=exception,
+                        )
 
-                if exception is not None:
-                    raise StopUser()
+                    if exception is not None:
+                        raise StopUser()
 
             paths: List[str] = key.split('.')
             variable = paths.pop(0)
@@ -140,15 +143,15 @@ def _objectify(testdata: Dict[str, Any]) -> Dict[str, Any]:
     return testdata
 
 
-def create_context_variable(grizzly: GrizzlyContext, variable: str, value: str) -> Dict[str, Any]:
+def create_context_variable(grizzly: 'GrizzlyContext', variable: str, value: str) -> Dict[str, Any]:
     casted_value = resolve_variable(grizzly, value)
 
     variable = variable.lower().replace(' ', '_').replace('/', '.')
 
-    return transform({variable: casted_value}, objectify=False)
+    return transform(grizzly, {variable: casted_value}, objectify=False)
 
 
-def resolve_variable(grizzly: GrizzlyContext, value: str, guess_datatype: Optional[bool] = True) -> GrizzlyDictValueType:
+def resolve_variable(grizzly: 'GrizzlyContext', value: str, guess_datatype: Optional[bool] = True) -> GrizzlyDictValueType:
     if len(value) < 1:
         return value
