+++ grizzly/testdata/variables/csv_row.py
@@ -61,7 +61,8 @@ from random import randint
 
 from grizzly_extras.arguments import split_value, parse_arguments
 
-from ...types import bool_typed, AtomicVariable
+from ...types import bool_typed
+from . import AtomicVariable
 
 
 def atomiccsvrow__base_type__(value: str) -> str:
