+++ grizzly/testdata/variables/servicebus.py
@@ -92,9 +92,10 @@ from grizzly_extras.async_message import AsyncMessageContext, AsyncMessageReques
 from grizzly_extras.arguments import split_value, parse_arguments, get_unsupported_arguments
 from grizzly_extras.transformer import TransformerContentType
 
-from ...types import AtomicVariable, RequestType, bool_typed
+from ...types import RequestType, bool_typed
 from ...context import GrizzlyContext
 from ..utils import resolve_variable
+from . import AtomicVariable
 
 
 def atomicservicebus__base_type__(value: str) -> str:
