+++ tests/helpers.py
@@ -10,7 +10,8 @@ from locust import task
 from locust.event import EventHook
 
 from grizzly.users.base import GrizzlyUser
-from grizzly.types import GrizzlyResponse, RequestMethod, Message, Environment, AtomicVariable
+from grizzly.types import GrizzlyResponse, RequestMethod, Message, Environment
+from grizzly.testdata.variables import AtomicVariable
 from grizzly.tasks import RequestTask
 from grizzly.scenarios import GrizzlyScenario
