+++ tests/test_grizzly/test_types.py
@@ -1,13 +1,6 @@
-from os import path, environ, mkdir
-from shutil import rmtree
-
 import pytest
 
-from _pytest.tmpdir import TempPathFactory
-
-from grizzly.types import RequestType, RequestDirection, RequestMethod, bool_typed, int_rounded_float_typed, AtomicVariable, GrizzlyDict
-
-from ..fixtures import AtomicVariableCleanupFixture
+from grizzly.types import RequestType, RequestDirection, RequestMethod, bool_typed, int_rounded_float_typed
 
 
 class TestRequestDirection:
@@ -96,311 +89,3 @@ def test_int_rounded_float_typed() -> None:
 
     with pytest.raises(ValueError):
         int_rounded_float_typed('1,5')
-
-
-class TestGrizzlyDict:
-    def test_static_str(self) -> None:
-        t = GrizzlyDict()
-
-        t['test1'] = 'hallo'
-        assert isinstance(t['test1'], str)
-
-        t['test7'] = '"true"'
-        assert isinstance(t['test7'], str)
-
-        t['test4'] = "'1337'"
-        assert isinstance(t['test4'], str)
-
-        t['test5'] = '"1337"'
-        assert isinstance(t['test5'], str)
-        assert t['test5'] == '1337'
-
-        t['test6'] = '"True"'
-        assert isinstance(t['test6'], str)
-        assert t['test6'] == 'True'
-
-        t['test7'] = '00004302'
-        assert isinstance(t['test7'], str)
-        assert t['test7'] == '00004302'
-
-        t['test8'] = '02002-00000'
-        assert isinstance(t['test8'], str)
-        assert t['test8'] == '02002-00000'
-
-    def test_static_float(self) -> None:
-        t = GrizzlyDict()
-
-        t['test2'] = 1.337
-        assert isinstance(t['test2'], float)
-
-        t['test2.1'] = -1.337
-        assert isinstance(t['test2.1'], float)
-
-        t['test2.2'] = '1.337'
-        assert isinstance(t['test2.2'], float)
-        assert t['test2.2'] == 1.337
-
-        t['test2.3'] = '-1.337'
-        assert isinstance(t['test2.3'], float)
-        assert t['test2.3'] == -1.337
-
-        t['test2.4'] = '0.01'
-        assert isinstance(t['test2.4'], float)
-        assert t['test2.4'] == 0.01
-
-    def test_static_int(self) -> None:
-        t = GrizzlyDict()
-        t['test3'] = 1337
-        assert isinstance(t['test3'], int)
-
-        t['test3.1'] = 1337
-        assert isinstance(t['test3.1'], int)
-
-        t['test3.2'] = '1337'
-        assert isinstance(t['test3.2'], int)
-        assert t['test3.2'] == 1337
-
-        t['test3.3'] = '-1337'
-        assert isinstance(t['test3.3'], int)
-        assert t['test3.3'] == -1337
-
-    def test_static_bool(self) -> None:
-        t: GrizzlyDict = GrizzlyDict()
-
-        t['test6'] = True
-        assert isinstance(t['test6'], bool)
-
-        t['test8'] = 'True'
-        assert isinstance(t['test8'], bool)
-
-        t['test9'] = 'False'
-        assert isinstance(t['test9'], bool)
-
-        t['test10'] = 'true'
-        assert isinstance(t['test10'], bool)
-        assert t.__getitem__('test10') is True
-
-        t['test11'] = 'FaLsE'
-        assert isinstance(t['test11'], bool)
-        assert t.__getitem__('test11') is False
-
-    def test_AtomicIntegerIncrementer(self, cleanup: AtomicVariableCleanupFixture) -> None:
-        try:
-            t = GrizzlyDict()
-            t['AtomicIntegerIncrementer.test1'] = 1337
-            assert isinstance(t['AtomicIntegerIncrementer.test1'], str)
-
-            t['AtomicIntegerIncrementer.test2'] = -1337
-            assert isinstance(t['AtomicIntegerIncrementer.test2'], str)
-
-            t['AtomicIntegerIncrementer.test3'] = '1337'
-            assert isinstance(t['AtomicIntegerIncrementer.test3'], str)
-            assert t['AtomicIntegerIncrementer.test3'] == '1337'
-
-            t['AtomicIntegerIncrementer.test4'] = '-1337'
-            assert isinstance(t['AtomicIntegerIncrementer.test4'], str)
-            assert t['AtomicIntegerIncrementer.test4'] == '-1337'
-
-            t['AtomicIntegerIncrementer.test5'] = '1.337'
-            assert t['AtomicIntegerIncrementer.test5'] == '1'
-
-            with pytest.raises(ValueError):
-                t['AtomicIntegerIncrementer.test6'] = 'hello'
-
-            t['AtomicIntegerIncrementer.test7'] = '1337 | step=10'
-            assert t['AtomicIntegerIncrementer.test7'] == '1337 | step=10'
-
-            t['AtomicIntegerIncrementer.test8'] = '1337|step=1'
-            assert t['AtomicIntegerIncrementer.test8'] == '1337 | step=1'
-
-            t['AtomicIntegerIncrementer.test9'] = '-1337|step=-10'
-            assert isinstance(t['AtomicIntegerIncrementer.test9'], str)
-            assert t['AtomicIntegerIncrementer.test9'] == '-1337 | step=-10'
-
-            t['AtomicIntegerIncrementer.test10'] = '1.337 | step=-1'
-            assert t['AtomicIntegerIncrementer.test10'] == '1 | step=-1'
-        finally:
-            cleanup()
-
-    def test_AtomicDirectoryContents(self, cleanup: AtomicVariableCleanupFixture, tmp_path_factory: TempPathFactory) -> None:
-        test_context = tmp_path_factory.mktemp('test_context') / 'requests'
-        test_context.mkdir()
-        test_context_root = path.dirname(test_context)
-        environ['GRIZZLY_CONTEXT_ROOT'] = test_context_root
-
-        try:
-            t = GrizzlyDict()
-
-            with pytest.raises(ValueError):
-                t['AtomicDirectoryContents.test1'] = 'doesnotexist/'
-
-            with open(path.join(test_context, 'notadirectory'), 'w') as fd:
-                fd.write('test')
-                fd.flush()
-
-            with pytest.raises(ValueError):
-                t['AtomicDirectoryContents.test2'] = 'notadirectory'
-
-            mkdir(path.join(test_context, 'adirectory'))
-
-            t['AtomicDirectoryContents.test3'] = 'adirectory'
-
-            with pytest.raises(ValueError):
-                t['AtomicDirectoryContents.test4'] = 'adirectory | repeat=asdf'
-
-            with pytest.raises(ValueError):
-                t['AtomicDirectoryContents.test4'] = 'adirectory | repeat=True, prefix="test-"'
-
-            t['AtomicDirectoryContents.test4'] = 'adirectory | repeat=True'
-
-            t['AtomicDirectoryContents.test5'] = 'adirectory|repeat=True'
-            assert t['AtomicDirectoryContents.test5'] == 'adirectory | repeat=True'
-
-            t['AtomicDirectoryContents.test6'] = 'adirectory| random=True'
-            assert t['AtomicDirectoryContents.test6'] == 'adirectory | random=True'
-
-            t['AtomicDirectoryContents.test7'] = 'adirectory|repeat=True, random=True'
-            assert t['AtomicDirectoryContents.test7'] == 'adirectory | repeat=True, random=True'
-        finally:
-            try:
-                del environ['GRIZZLY_CONTEXT_ROOT']
-            except:
-                pass
-
-            rmtree(test_context_root)
-            cleanup()
-
-    def test_AtomicCsvRow(self, cleanup: AtomicVariableCleanupFixture, tmp_path_factory: TempPathFactory) -> None:
-        test_context = tmp_path_factory.mktemp('test_context') / 'requests'
-        test_context.mkdir()
-        test_context_root = path.dirname(test_context)
-        environ['GRIZZLY_CONTEXT_ROOT'] = test_context_root
-
-        with open(path.join(test_context, 'test.csv'), 'w') as fd:
-            fd.write('header1,header2\n')
-            fd.write('value1,value2\n')
-            fd.flush()
-
-        try:
-            t = GrizzlyDict()
-
-            with pytest.raises(ValueError):
-                t['AtomicCsvRow.test'] = 'doesnotexist.csv'
-
-            t['AtomicCsvRow.test'] = 'test.csv'
-
-            with pytest.raises(ValueError):
-                t['AtomicCsvRow.test2'] = 'test.csv | repeat=asdf'
-
-            with pytest.raises(ValueError):
-                t['AtomicCsvRow.test2'] = 'test.csv | repeat=True, suffix=True'
-
-            t['AtomicCsvRow.test2'] = 'test.csv|repeat=True'
-            assert t['AtomicCsvRow.test2'] == 'test.csv | repeat=True'
-        finally:
-            try:
-                del environ['GRIZZLY_CONTEXT_ROOT']
-            except:
-                pass
-
-            rmtree(test_context_root)
-            cleanup()
-
-    def test_AtomicDate(self, cleanup: AtomicVariableCleanupFixture) -> None:
-        try:
-            t = GrizzlyDict()
-            with pytest.raises(ValueError):
-                t['AtomicDate.test1'] = 1337
-
-            with pytest.raises(ValueError):
-                t['AtomicDate.test6'] = 'hello'
-
-            t['AtomicDate.test2'] = '2021-03-29'
-            assert isinstance(t['AtomicDate.test2'], str)
-            assert t['AtomicDate.test2'] == '2021-03-29'
-
-            t['AtomicDate.test3'] = '2021-03-29 16:43:49'
-            assert isinstance(t['AtomicDate.test3'], str)
-            assert t['AtomicDate.test3'] == '2021-03-29 16:43:49'
-
-            t['AtomicDate.test4'] = 'now|format="%Y-%m-%d"'
-            assert isinstance(t['AtomicDate.test4'], str)
-            assert t['AtomicDate.test4'] == 'now | format="%Y-%m-%d"'
-
-            with pytest.raises(ValueError):
-                t['AtomicDate.test5'] = 'asdf|format="%Y-%m-%d"'
-
-            with pytest.raises(ValueError):
-                t['AtomicDate.test6'] = 'now|'
-
-            t['AtomicDate.test7'] = 'now | format="%Y-%m-%dT%H:%M:%S.000Z"'
-        finally:
-            cleanup()
-
-    def test_AtomicRandomInteger(self, cleanup: AtomicVariableCleanupFixture) -> None:
-        try:
-            t = GrizzlyDict()
-
-            with pytest.raises(ValueError):
-                t['AtomicRandomInteger.test1'] = '10'
-
-            with pytest.raises(ValueError):
-                t['AtomicRandomInteger.test2'] = '1.17..5.0'
-
-            with pytest.raises(ValueError):
-                t['AtomicRandomInteger.test5'] = '1.0..3.5'
-
-            with pytest.raises(ValueError):
-                t['AtomicRandomInteger.test3'] = '100..10'
-
-            t['AtomicRandomInteger.test4'] = '1..10'
-            assert isinstance(t['AtomicRandomInteger.test4'], str)
-            assert t['AtomicRandomInteger.test4'] == '1..10'
-        finally:
-            cleanup()
-
-
-class TestAtomicVariable:
-    def test_dont_instantiate(self, cleanup: AtomicVariableCleanupFixture) -> None:
-        try:
-            with pytest.raises(TypeError):
-                AtomicVariable('dummy')
-
-            with pytest.raises(TypeError):
-                AtomicVariable[int]('dummy')
-        finally:
-            cleanup()
-
-    def test_get(self) -> None:
-        with pytest.raises(ValueError) as ve:
-            AtomicVariable.get()
-        assert 'is not instantiated' in str(ve)
-
-    def test_destroy(self) -> None:
-        with pytest.raises(ValueError) as ve:
-            AtomicVariable.destroy()
-        assert 'is not instantiated' in str(ve)
-
-    def test_clear(self) -> None:
-        with pytest.raises(ValueError) as ve:
-            AtomicVariable.clear()
-        assert 'is not instantiated' in str(ve)
-
-    def test___getitem___and___setitem__(self) -> None:
-        class AtomicTestVariable(AtomicVariable):
-            pass
-
-        t = AtomicTestVariable('hello', 'value')
-
-        with pytest.raises(AttributeError) as ae:
-            t['foo']
-        assert "'AtomicTestVariable' object has no attribute 'foo'" == str(ae.value)
-
-        assert t['hello'] == 'value'
-
-        with pytest.raises(AttributeError) as ae:
-            t['foo'] = 'bar'
-        assert "'AtomicTestVariable' object has no attribute 'foo'" == str(ae.value)
-
-        t['hello'] = 'bar'
-        assert t['hello'] == 'bar'
